+++ CHANGELOG.md
@@ -722,3 +722,9 @@ This major version introduces the concept of `Engines` which handle the chunking
   - New util `ReaderWRG`, reading WRG files. Thanks @HassanKassem
 
 **Full Changelog**: [https://github.com/FraunhoferIWES/foxes/commits/v1.2.4](https://github.com/FraunhoferIWES/foxes/commits/v1.2.4)
+
+## v1.2.5
+
+
+
+**Full Changelog**: [https://github.com/FraunhoferIWES/foxes/commits/v1.2.5](https://github.com/FraunhoferIWES/foxes/commits/v1.2.5)